Added reassoc to make command-lists right-assoc.
The original syntax tree used right recursion for
commands strung together by ';', '&', and '\n' (inside
of braces).  When changing the syntax to left-recursive,
the association order of tree2(';',a,b) changed as well.
This confused the interpreter when looking for case
statements inside switch().  This fix changes all consumers
of `line` and `body` syntax trees to make an in-place
change of associativity.